+++ components/toast/method.js
@@ -1,40 +1,59 @@
 // eslint-disable-next-line
 import Toast from './Toast'
 // eslint-disable-next-line
-import { h } from 'hyperapp'
-import { createApp } from '../_util'
+import { h, app } from 'hyperapp'
+import { createElement } from '../_util'
 
 const DURATION = 1500
 
-function text (msg, duration = DURATION) {
-  createApp((destroy) => {
-    return [
-      {
-        show: true,
-        duration
-      },
-      {
-        close: () => {
-          return { show: false }
-        },
-        schedule: () => (state, actions) => {
-          setTimeout(actions.close, state.duration)
-        }
-      },
-      (state, actions) => {
-        return (state.show &&
-          <Toast
-            oncreate={actions.schedule}
-            ondestroy={destroy}
-          >
-            {msg}
-          </Toast>
-        )
+function create (msg, duration = DURATION, onClose, mask, onClick) {
+  const { div, remove } = createElement()
+  return app(
+    { show: true },
+    {
+      close: () => {
+        return { show: false }
       }
-    ]
-  })
+    },
+    (state, actions) => {
+      return (
+        <Toast
+          show={state.show}
+          mask={mask}
+          onOpen={() => {
+            if (duration !== 0) {
+              setTimeout(actions.close, duration)
+            }
+          }}
+          onToastClick={onClick}
+          onClose={() => {
+            onClose && onClose()
+            remove()
+          }}
+        >
+          {msg}
+        </Toast>
+      )
+    },
+    div
+  )
+}
+
+let _close
+
+function text (msg, duration, onClose, mask, onClick) {
+  hide()
+  _close = create(msg, duration, onClose, mask, onClick).close
+}
+
+function hide () {
+  if (_close) {
+    _close()
+  }
+  _close = null
 }
 
 export default {
-  text
+  text,
+  hide
 }